requirements.txt
streamlit
langchain
langchain-community
langchain_google_genai
faiss-cpu
PyPDF2
pandas
python-docx
openpyxl
psycopg2-binary
bcrypt

